Join adaptive pooling code
