+++ laia/nn/__init__.py
@@ -1,5 +1,4 @@
-from laia.nn.adaptive_avgpool_2d import AdaptiveAvgPool2d
-from laia.nn.adaptive_maxpool_2d import AdaptiveMaxPool2d
+from laia.nn.adaptive_pool_2d import AdaptiveAvgPool2d, AdaptiveMaxPool2d
 from laia.nn.image_pooling_sequencer import ImagePoolingSequencer
 from laia.nn.image_to_sequence import ImageToSequence
 from laia.nn.mask_image_from_size import MaskImageFromSize
